pyproject.toml
[project]
name = "playnet"
version = "0.1.0"
description = "Game genre classfication from ingame screeshot images"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "ipykernel>=7.1.0",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "fastapi>=0.101.0",
    "requests>=2.31.0",
    "Pillow>=10.0.1",
    "tqdm>=4.66.1",
    "streamlit>=1.25.0",
]

[tool.basedpyright]
typeCheckingMode = "basic"
reportAttributeAccessIssue = false
